refactor(phase4): extract policy_plan_tokens from codegen.rs
Move the policy-plan generators, validation helpers, insert/bind
utilities and UpdatePlan out of codegen.rs (7,252 → 5,625 lines) into
codegen/policy_plan_tokens.rs (1,661 lines).

The new module is a child of codegen, so it accesses the handful of
parent helpers it needs via explicit `use super::` imports. UpdatePlan
fields are pub(super) so the parent's resource_impl_tokens can pattern-
match on the plan struct.

codegen.rs now owns only the four public entry-points, resource-struct
emission (resource_struct_tokens and its validation/typed-object helpers),
and the resource-impl body (resource_impl_tokens with SQL/handler code).
